pyproject.toml
[tool.poetry]
name = "raindrop-video-summarizer"
version = "0.1.0"
description = "Raindrop.io Video Summarizer with LLM"
authors = ["Antigravity"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.5"
python-dotenv = "^1.2.1"
yt-dlp = "^2025.01.12"
google-generativeai = "^0.8.4"
pydantic = "^2.10.0"
loguru = "^0.7.2"
tenacity = "^8.5.0"
markdown = "^3.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
